submission/requirements.txt
torch>=1.13.0
torchvision>=0.14.0
timm>=0.9.2
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
Pillow>=9.0.0

